refactor: move out fixture
